Expose Subscription properties
